+++ Libplanet.Tests/Net/SwarmTest.TrustedStateCompleter.cs
@@ -1,0 +1,90 @@
+using System;
+using System.Collections.Immutable;
+using System.Linq;
+using System.Security.Cryptography;
+using System.Threading.Tasks;
+using Bencodex.Types;
+using Libplanet.Blockchain;
+using Libplanet.Net;
+using Libplanet.Store;
+using Libplanet.Tests.Blockchain;
+using Libplanet.Tests.Common.Action;
+using Serilog.Events;
+using Xunit;
+
+namespace Libplanet.Tests.Net
+{
+    public partial class SwarmTest
+    {
+        [Fact]
+        public async Task TrustedStateCompleter()
+        {
+            StateCompleter<DumbAction> noComplete = StateCompleters<DumbAction>.Reject;
+
+            // The chain with incomplete states
+            (_, Address[] addresses, BlockChain<DumbAction> incompleteChain) =
+                BlockChainTest.MakeIncompleteBlockStates(new DefaultStore(null));
+            Swarm<DumbAction> incompleteSwarm = CreateSwarm(incompleteChain);
+            Assert.Throws<IncompleteBlockStatesException>(() =>
+                incompleteChain.GetState(addresses[0], stateCompleter: noComplete)
+            );
+
+            // The chain with complete states
+            var store = new DefaultStore(null);
+            BlockChain<DumbAction> completeChain =
+                new BlockChain<DumbAction>(incompleteChain.Policy, store, incompleteChain.Genesis);
+            Swarm<DumbAction> completeSwarm = CreateSwarm(completeChain);
+            foreach (HashDigest<SHA256> blockHash in incompleteChain.BlockHashes.Skip(1))
+            {
+                completeChain.Append(incompleteChain[blockHash]);
+            }
+
+            _logger.CompareBothChains(
+                LogEventLevel.Debug,
+                "incomplete",
+                incompleteChain,
+                "complete",
+                completeChain
+            );
+            IValue completeState = completeChain.GetState(
+                addresses[0],
+                stateCompleter: StateCompleters<DumbAction>.Reject
+            );
+            Assert.Equal((Text)"0,5", completeState);
+
+            // Network
+            await StartAsync(completeSwarm);
+            await StartAsync(incompleteSwarm);
+            await incompleteSwarm.AddPeersAsync(new[] { completeSwarm.AsPeer }, null);
+            Assert.Throws<IncompleteBlockStatesException>(() =>
+                incompleteChain.GetState(addresses[0], stateCompleter: noComplete)
+            );
+
+            // Empty trustedStateValidators disallow to reuse others states:
+            StateCompleterSet<DumbAction> noTrust =
+                await incompleteSwarm.GetTrustedStateCompleterAsync(
+                    ImmutableHashSet<Address>.Empty,
+                    TimeSpan.FromSeconds(10),
+                    fallbackCompleterSet: StateCompleterSet<DumbAction>.Reject
+                );
+            Assert.Throws<IncompleteBlockStatesException>(() =>
+                incompleteChain.GetState(addresses[0], stateCompleter: noTrust.StateCompleter)
+            );
+
+            // If any trustStateValidators have block states reuse them:
+            StateCompleterSet<DumbAction> scs = await incompleteSwarm.GetTrustedStateCompleterAsync(
+                ImmutableHashSet.Create(completeSwarm.Address),
+                TimeSpan.FromSeconds(10),
+                fallbackCompleterSet: StateCompleterSet<DumbAction>.Reject
+            );
+            Assert.Equal(
+                (Text)"0,5",
+                incompleteChain.GetState(addresses[0], stateCompleter: scs.StateCompleter)
+            );
+            Assert.Equal(
+                (Text)"0,5",
+                incompleteChain.GetState(addresses[0], stateCompleter: noComplete)
+            );
+        }
+    }
+}